Modified: extracted pokemon detail into separate component
